# Enable schema caching for tests