--- a/credential/manifest/schema_test.go
+++ b/credential/manifest/schema_test.go
@@ -13,11 +13,15 @@ import (
 func TestMain(m *testing.M) {
 	localSchemas, err := schema.GetAllLocalSchemas()
 	if err != nil {
+		println(err)
 		os.Exit(1)
 	}
-	if _, err = schema.NewCachingLoader(localSchemas); err != nil {
+	loader, err := schema.NewCachingLoader(localSchemas)
+	if err != nil {
+		println(err)
 		os.Exit(1)
 	}
+	loader.EnableHTTPCache()
 	os.Exit(m.Run())
 }